stats handler

+++ database/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, String, ForeignKey, Float
 from sqlalchemy.orm import declarative_base, relationship
 
 Base = declarative_base()
@@ -7,5 +7,5 @@ class DBUser(Base):
     __tablename__ = "users"
     id = Column(Integer, primary_key=True, index=True)
     username = Column(String, index=True)
-    height = Column(Integer, index=True)
-    weight = Column(Integer, index=True)
+    height = Column(Float, index=True)
+    weight = Column(Float, index=True)
